# v0.19.1 | Performance optimisations on v0.19.0