--- a/lib/blueprinter/extractors/auto_extractor.rb
+++ b/lib/blueprinter/extractors/auto_extractor.rb
@@ -5,18 +5,24 @@ module Blueprinter
       @public_send_extractor = PublicSendExtractor.new
       @block_extractor = BlockExtractor.new
       @datetime_formatter = DateTimeFormatter.new
+      @config_datetime_format = Blueprinter.configuration.datetime_format
+      @config_field_default = Blueprinter.configuration.field_default
     end
 
     def extract(field_name, object, local_options, options = {})
-      extraction = extractor(object, options).extract(field_name, object, local_options, options)
-      value = @datetime_formatter.format(extraction, options)
+      value = extractor(object, options).extract(field_name, object, local_options, options)
+      value = @datetime_formatter.format(value, options) if format_datetime?(options)
       value.nil? ? default_value(options) : value
     end
 
     private
 
+    def format_datetime?(options)
+      options.key?(:datetime_format) || @config_datetime_format
+    end
+
     def default_value(field_options)
-      field_options.key?(:default) ? field_options.fetch(:default) : Blueprinter.configuration.field_default
+      field_options.key?(:default) ? field_options.fetch(:default) : @config_field_default
     end
 
     def extractor(object, options)