--- a/lib/blueprinter/helpers/base_helpers.rb
+++ b/lib/blueprinter/helpers/base_helpers.rb
@@ -20,15 +20,19 @@ module Blueprinter
       end
 
       def prepare_data(object, view_name, local_options)
+        view_fields = view_collection.fields_for(view_name)
+        view_transformers = view_collection.transformers(view_name)
         if array_like?(object)
           object.map do |obj|
             object_to_hash(obj,
-                          view_name: view_name,
+                          view_fields: view_fields,
+                          view_transformers: view_transformers,
                           local_options: local_options)
           end
         else
           object_to_hash(object,
-                        view_name: view_name,
+                        view_fields: view_fields,
+                        view_transformers: view_transformers,
                         local_options: local_options)
         end
       end
@@ -43,12 +47,12 @@ module Blueprinter
         subclass.send(:view_collection).inherit(view_collection)
       end
 
-      def object_to_hash(object, view_name:, local_options:)
-        result_hash = view_collection.fields_for(view_name).each_with_object({}) do |field, hash|
+      def object_to_hash(object, view_fields:, view_transformers:, local_options:)
+        result_hash = view_fields.each_with_object({}) do |field, hash|
           next if field.skip?(field.name, object, local_options)
            hash[field.name] = field.extract(object, local_options)
         end
-        view_collection.transformers(view_name).each do |transformer|
+        view_transformers.each do |transformer|
            transformer.transform(result_hash,object,local_options)
         end
         result_hash