### Benchmark comparison:

Output on Ruby 2.6.9, MacBook Pro (15-inch, 2018, 2.2 GHz 6-Core Intel Core i7)

[benchmark code](https://gist.github.com/midhun-thimmapuram/f6b45744e77f01750db64eb5f7a7a61e)

_blueprinter 0.19.0 (procore)_

```
Rehearsal -----------------------------------------------
blueprinter   0.276568   0.005310   0.281878 (  0.283335)
-------------------------------------- total: 0.281878sec

                  user     system      total        real
blueprinter   0.255435   0.005273   0.260708 (  0.264110)
Warming up --------------------------------------
         blueprinter     1.000  i/100ms
Calculating -------------------------------------
         blueprinter      3.644  (± 3.4%) i/s -     37.000  in  10.300638s
                   with 95.0% confidence
Calculating -------------------------------------
         blueprinter    35.117M memsize (     0.000  retained)
                       279.762k objects (     0.000  retained)
                         0.000  strings (     0.000  retained)
```
_With current PR optimisations (0.19.1)_
```
Rehearsal -----------------------------------------------
blueprinter   0.177554   0.002541   0.180095 (  0.180207)
-------------------------------------- total: 0.180095sec

                  user     system      total        real
blueprinter   0.162315   0.000225   0.162540 (  0.162751)
Warming up --------------------------------------
         blueprinter     1.000  i/100ms
Calculating -------------------------------------
         blueprinter      5.852  (± 1.8%) i/s -     59.000  in  10.128980s
                   with 95.0% confidence
Calculating -------------------------------------
         blueprinter    23.839M memsize (     0.000  retained)
                       159.889k objects (     0.000  retained)
                         0.000  strings (     0.000  retained)
```